Cleanup in serializer tests: Moved types to matching files. No changes made.
